requirements.txt
# Python
python>=3.8

# Deep Learning
torch>=1.10
torchvision>=0.11

# UI / Apps
streamlit>=1.25
gradio>=4.0

# Numerical & Scientific
numpy>=1.21
scipy>=1.7
pandas>=1.5

# Image Processing
opencv-python>=4.5
scikit-image>=0.19
Pillow>=8.4
matplotlib>=3.5

# Geospatial
rasterio>=1.3
geopandas>=0.12
shapely>=2.0
affine>=2.4

# Machine Learning
scikit-learn>=1.2
xgboost>=1.7
joblib>=1.2

# Utilities
tqdm>=4.64
einops>=0.6

# Optional (SR / Research)
albumentations>=1.3
lpips>=0.1.4

# Notebook Support
jupyter
ipykernel
